refactor: move file io types to io directory
