+++ .github/cicd/core/io/Directory.ts
@@ -1,6 +1,6 @@
-import { Guard } from "./Guard.ts";
-import { Path } from "./Path.ts";
-import { Utils } from "./Utils.ts";
+import { Guard } from "../Guard.ts";
+import { Path } from "io/Path.ts";
+import { Utils } from "../Utils.ts";
 
 /**
  * Provides directory related operations.
