+++ .github/cicd/core/io/File.ts
@@ -1,6 +1,6 @@
-import { Guard } from "./Guard.ts";
+import { Guard } from "../Guard.ts";
 import { Path } from "./Path.ts";
-import { Utils } from "./Utils.ts";
+import { Utils } from "../Utils.ts";
 
 /**
  * Provides file related operations.
